FEAT: Added Logo to Website
added company logo to the website. Insured the change was made to aboutus
and contactus HTML pages as well. Created the additional pitures that will
be add into the site in a future change to the code base. Added the
Picture Source Zip to the repo. Add the PDF Slides for the Imgaes and
media bootstrap lectures.